+++ src/main/kotlin/org/dreamexposure/ticketbird/command/TopicCommand.kt
@@ -3,28 +3,16 @@ package org.dreamexposure.ticketbird.command
 import discord4j.core.event.domain.interaction.ChatInputInteractionEvent
 import discord4j.core.`object`.command.ApplicationCommandInteractionOption
 import discord4j.core.`object`.command.ApplicationCommandInteractionOptionValue
-import discord4j.core.`object`.entity.Message
-import discord4j.core.`object`.entity.channel.TextChannel
-import kotlinx.coroutines.reactor.awaitSingleOrNull
-import org.dreamexposure.ticketbird.business.LocaleService
-import org.dreamexposure.ticketbird.business.ProjectService
-import org.dreamexposure.ticketbird.business.TicketService
-import org.dreamexposure.ticketbird.config.Config
-import org.dreamexposure.ticketbird.extensions.asSeconds
-import org.dreamexposure.ticketbird.extensions.discord4j.deleteFollowupDelayed
+import org.dreamexposure.ticketbird.business.InteractionService
 import org.dreamexposure.ticketbird.`object`.GuildSettings
 import org.springframework.stereotype.Component
 
 @Component
 class TopicCommand(
-    private val ticketService: TicketService,
-    private val projectService: ProjectService,
-    private val localeService: LocaleService,
+    private val interactionService: InteractionService,
 ): SlashCommand {
     override val name = "topic"
     override val ephemeral = true
-
-    private val messageDeleteSeconds = Config.TIMING_MESSAGE_DELETE_GENERIC_SECONDS.getLong().asSeconds()
 
     override suspend fun handle(event: ChatInputInteractionEvent, settings: GuildSettings) {
         val topicId = event.getOption("topic")
@@ -32,42 +20,7 @@ class TopicCommand(
             .map(ApplicationCommandInteractionOptionValue::asString)
             .map(String::toLong)
             .orElse(-1)
-        val ticket = ticketService.getTicket(settings.guildId, event.interaction.channelId)
 
-        // Handle if not in a ticket channel
-        if (ticket == null) {
-            event.createFollowup(localeService.getString(settings.locale, "command.topic.not-ticket"))
-                .withEphemeral(ephemeral)
-                .map(Message::getId)
-                .flatMap { event.deleteFollowupDelayed(it, messageDeleteSeconds) }
-                .awaitSingleOrNull()
-            return
-        }
-
-        val project = projectService.getProject(settings.guildId, topicId)
-
-        if (project == null) {
-            event.createFollowup(localeService.getString(settings.locale, "command.topic.not-found"))
-                .withEphemeral(ephemeral)
-                .map(Message::getId)
-                .flatMap { event.deleteFollowupDelayed(it, messageDeleteSeconds) }
-                .awaitSingleOrNull()
-            return
-        }
-
-        // Update
-        event.interaction.channel
-            .ofType(TextChannel::class.java)
-            .flatMap { it.edit().withName("${project.prefix}-ticket-${ticket.number}") }
-            .awaitSingleOrNull()
-
-        ticket.project = project.name
-        ticketService.updateTicket(ticket)
-
-        event.createFollowup(localeService.getString(settings.locale, "command.topic.success", project.name))
-            .withEphemeral(ephemeral)
-            .map(Message::getId)
-            .flatMap { event.deleteFollowupDelayed(it, messageDeleteSeconds) }
-            .awaitSingleOrNull()
+        interactionService.changeTopicViaCommand(topicId, ephemeral, event, settings)
     }
 }
